--- tests/TestVerifiers.Library/CSharpCodeFixVerifier`2+Test.cs
@@ -1,30 +1,0 @@
-﻿namespace Treasure.Analyzers.MemberOrder.Test.Verifiers;
-
-using Microsoft.CodeAnalysis.CodeFixes;
-using Microsoft.CodeAnalysis.CSharp.Testing;
-using Microsoft.CodeAnalysis.Diagnostics;
-using Microsoft.CodeAnalysis.Testing.Verifiers;
-
-using Treasure.Analyzers.TestVerifiers;
-
-public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
-    where TAnalyzer : DiagnosticAnalyzer, new()
-    where TCodeFix : CodeFixProvider, new()
-{
-    internal sealed class Test : CSharpCodeFixTest<TAnalyzer, TCodeFix, MSTestVerifier>
-    {
-        public Test()
-        {
-            this.SolutionTransforms.Add((solution, projectId) =>
-            {
-                Microsoft.CodeAnalysis.CompilationOptions? compilationOptions = solution.GetProject(projectId)?.CompilationOptions
-                    ?? throw new InvalidOperationException("Unable to get projects' compilation options");
-                compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
-                    compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
-                solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
-
-                return solution;
-            });
-        }
-    }
-}
