+++ tests/TestVerifiers.Library/CSharpCodeFixVerifier`2.cs
@@ -9,29 +9,30 @@ using Microsoft.CodeAnalysis.CodeFixes;
 using Microsoft.CodeAnalysis.CSharp.Testing;
 using Microsoft.CodeAnalysis.Diagnostics;
 using Microsoft.CodeAnalysis.Testing;
-using Microsoft.CodeAnalysis.Testing.Verifiers;
 
-[SuppressMessage("Design", "CA1000:Do not declare static members on generic types", Justification = "<Pending>")]
-public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
+using Treasure.Analyzers.TestVerifiers;
+
+[SuppressMessage("Design", "CA1000:Do not declare static members on generic types")]
+public static class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
     where TAnalyzer : DiagnosticAnalyzer, new()
     where TCodeFix : CodeFixProvider, new()
 {
     /// <inheritdoc cref="CodeFixVerifier{TAnalyzer, TCodeFix, TTest, TVerifier}.Diagnostic()"/>
     public static DiagnosticResult Diagnostic()
-        => CSharpCodeFixVerifier<TAnalyzer, TCodeFix, MSTestVerifier>.Diagnostic();
+        => CSharpCodeFixVerifier<TAnalyzer, TCodeFix, DefaultVerifier>.Diagnostic();
 
     /// <inheritdoc cref="CodeFixVerifier{TAnalyzer, TCodeFix, TTest, TVerifier}.Diagnostic(string)"/>
     public static DiagnosticResult Diagnostic(string diagnosticId)
-        => CSharpCodeFixVerifier<TAnalyzer, TCodeFix, MSTestVerifier>.Diagnostic(diagnosticId);
+        => CSharpCodeFixVerifier<TAnalyzer, TCodeFix, DefaultVerifier>.Diagnostic(diagnosticId);
 
     /// <inheritdoc cref="CodeFixVerifier{TAnalyzer, TCodeFix, TTest, TVerifier}.Diagnostic(DiagnosticDescriptor)"/>
     public static DiagnosticResult Diagnostic(DiagnosticDescriptor descriptor)
-        => CSharpCodeFixVerifier<TAnalyzer, TCodeFix, MSTestVerifier>.Diagnostic(descriptor);
+        => CSharpCodeFixVerifier<TAnalyzer, TCodeFix, DefaultVerifier>.Diagnostic(descriptor);
 
     /// <inheritdoc cref="CodeFixVerifier{TAnalyzer, TCodeFix, TTest, TVerifier}.VerifyAnalyzerAsync(string, DiagnosticResult[])"/>
     public static async Task VerifyAnalyzerAsync(string source, params DiagnosticResult[] expected)
     {
-        CSharpCodeFixVerifier<TAnalyzer, TCodeFix>.Test test = new()
+        Test test = new()
         {
             TestCode = source,
         };
@@ -51,7 +52,7 @@ public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
     /// <inheritdoc cref="CodeFixVerifier{TAnalyzer, TCodeFix, TTest, TVerifier}.VerifyCodeFixAsync(string, DiagnosticResult[], string)"/>
     public static async Task VerifyCodeFixAsync(string source, DiagnosticResult[] expected, string fixedSource)
     {
-        CSharpCodeFixVerifier<TAnalyzer, TCodeFix>.Test test = new()
+        Test test = new()
         {
             TestCode = source,
             FixedCode = fixedSource,
@@ -59,5 +60,22 @@ public static partial class CSharpCodeFixVerifier<TAnalyzer, TCodeFix>
 
         test.ExpectedDiagnostics.AddRange(expected);
         await test.RunAsync(CancellationToken.None);
+    }
+
+    internal sealed class Test : CSharpCodeFixTest<TAnalyzer, TCodeFix, DefaultVerifier>
+    {
+        public Test()
+        {
+            this.SolutionTransforms.Add((solution, projectId) =>
+            {
+                CompilationOptions? compilationOptions = solution.GetProject(projectId)?.CompilationOptions
+                    ?? throw new InvalidOperationException("Unable to get projects' compilation options");
+                compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
+                    compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
+                solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
+
+                return solution;
+            });
+        }
     }
 }
