+++ tests/TestVerifiers.Library/CSharpCodeRefactoringVerifier`1.cs
@@ -5,10 +5,13 @@ using System.Threading;
 using System.Threading.Tasks;
 
 using Microsoft.CodeAnalysis.CodeRefactorings;
+using Microsoft.CodeAnalysis.CSharp.Testing;
 using Microsoft.CodeAnalysis.Testing;
 
-[SuppressMessage("Design", "CA1000:Do not declare static members on generic types", Justification = "<Pending>")]
-public static partial class CSharpCodeRefactoringVerifier<TCodeRefactoring>
+using Treasure.Analyzers.TestVerifiers;
+
+[SuppressMessage("Design", "CA1000:Do not declare static members on generic types")]
+public static class CSharpCodeRefactoringVerifier<TCodeRefactoring>
     where TCodeRefactoring : CodeRefactoringProvider, new()
 {
     /// <inheritdoc cref="CodeRefactoringVerifier{TCodeRefactoring, TTest, TVerifier}.VerifyRefactoringAsync(string, string)"/>
@@ -26,7 +29,7 @@ public static partial class CSharpCodeRefactoringVerifier<TCodeRefactoring>
     /// <inheritdoc cref="CodeRefactoringVerifier{TCodeRefactoring, TTest, TVerifier}.VerifyRefactoringAsync(string, DiagnosticResult[], string)"/>
     public static async Task VerifyRefactoringAsync(string source, DiagnosticResult[] expected, string fixedSource)
     {
-        CSharpCodeRefactoringVerifier<TCodeRefactoring>.Test test = new()
+        Test test = new()
         {
             TestCode = source,
             FixedCode = fixedSource,
@@ -34,5 +37,22 @@ public static partial class CSharpCodeRefactoringVerifier<TCodeRefactoring>
 
         test.ExpectedDiagnostics.AddRange(expected);
         await test.RunAsync(CancellationToken.None);
+    }
+
+    internal sealed class Test : CSharpCodeRefactoringTest<TCodeRefactoring, DefaultVerifier>
+    {
+        public Test()
+        {
+            this.SolutionTransforms.Add((solution, projectId) =>
+            {
+                Microsoft.CodeAnalysis.CompilationOptions? compilationOptions = solution.GetProject(projectId)?.CompilationOptions
+                    ?? throw new InvalidOperationException("Unable to get project's compilation options.");
+                compilationOptions = compilationOptions.WithSpecificDiagnosticOptions(
+                    compilationOptions.SpecificDiagnosticOptions.SetItems(CSharpVerifierHelper.NullableWarnings));
+                solution = solution.WithProjectCompilationOptions(projectId, compilationOptions);
+
+                return solution;
+            });
+        }
     }
 }
